added files used in 2016 analysis
'recode PNS for R' is a Stata do file to edit the main microdata file (which is Stata format) so that the response relating to 'prefer not to say' is recoded as 'PNS' instead of the missing value it has in the original data.  'for People Survey analysis' is an R script operating on the result of the Stata file.  'estimation' is an R file to be run after 'for People Survey analysis' which performs the regressions, then 'results database' puts all results into a dataframe, which is useful for 'graphs'